Empty Tensor Script
This commit adds an empty tensor file to the repository that will be used to upload to the microcontroller. The tensor will be filled with data at a later stage of the project, and this empty file ensures that the structure and format of the tensor are in place for future development. This commit is an important step towards the loading a trained machine learning algorithm on the microcontroller.